[dsp] rename _processors -> processors
