--- src/pygama/dsp/_processors/__init__.py
@@ -1,3 +1,0 @@
-"""
-Package documentation
-"""
