+++ src/pygama/dsp/processors/convolutions.py
