+++ src/pygama/dsp/processors/fixed_time_pickoff.py
