+++ src/pygama/dsp/processors/gaussian_filter1d.py
