+++ src/pygama/dsp/processors/get_multi_local_extrema.py
