+++ src/pygama/dsp/processors/linear_slope_fit.py
