+++ src/pygama/dsp/processors/log_check.py
