+++ src/pygama/dsp/processors/min_max.py
