+++ src/pygama/dsp/processors/moving_windows.py
