+++ src/pygama/dsp/processors/multi_t_filter.py
@@ -1,7 +1,7 @@
 import numpy as np
 from numba import guvectorize
 
-from pygama.dsp._processors.time_point_thresh import time_point_thresh
+from .time_point_thresh import time_point_thresh
 from pygama.dsp.errors import DSPFatal
