+++ src/pygama/dsp/processors/optimize.py
