+++ src/pygama/dsp/processors/param_lookup.py
