+++ src/pygama/dsp/processors/presum.py
