+++ src/pygama/dsp/processors/pulse_injector.py
