+++ src/pygama/dsp/processors/saturation.py
