+++ src/pygama/dsp/processors/soft_pileup_corr.py
