+++ src/pygama/dsp/processors/trap_filters.py
