+++ src/pygama/dsp/processors/windower.py
